@@ -1,14 +1,14 @@
 module github.com/favonia/cloudflare-ddns
 
-go 1.18
+go 1.19
 
 require (
-	github.com/cloudflare/cloudflare-go v0.45.0
+	github.com/cloudflare/cloudflare-go v0.50.0
 	github.com/golang/mock v1.6.0
 	github.com/patrickmn/go-cache v2.1.0+incompatible
 	github.com/robfig/cron/v3 v3.0.1
 	github.com/stretchr/testify v1.8.0
-	golang.org/x/net v0.0.0-20220526153639-5463443f8c37
+	golang.org/x/net v0.0.0-20220722155237-a158d28d115b
 	kernel.org/pub/linux/libs/security/libcap/cap v1.2.65
 )
 
@@ -17,6 +17,7 @@ require (
 	github.com/google/go-querystring v1.1.0 // indirect
 	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
 	github.com/hashicorp/go-retryablehttp v0.7.1 // indirect
+	github.com/kr/text v0.2.0 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/text v0.3.7 // indirect
