@@ -18,6 +18,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/favonia/cloudflare-ddns/internal/api"
+	"github.com/favonia/cloudflare-ddns/internal/domain"
 	"github.com/favonia/cloudflare-ddns/internal/ipnet"
 	"github.com/favonia/cloudflare-ddns/internal/mocks"
 	"github.com/favonia/cloudflare-ddns/internal/pp"
@@ -158,22 +159,24 @@ func TestNewInvalid(t *testing.T) {
 	require.Nil(t, h)
 }
 
-func mockZone(zoneName string, i int) *cloudflare.Zone {
+func mockZone(name string, i int, status string) *cloudflare.Zone {
 	return &cloudflare.Zone{ //nolint:exhaustruct
-		ID:     mockID(zoneName, i),
-		Name:   zoneName,
-		Status: "active",
+		ID:     mockID(name, i),
+		Name:   name,
+		Status: status,
 	}
 }
 
-func mockZonesResponse(zoneName string, numZones int) *cloudflare.ZonesResponse {
+func mockZonesResponse(zoneName string, zoneStatuses []string) *cloudflare.ZonesResponse {
+	numZones := len(zoneStatuses)
+
 	if numZones > 50 {
 		panic("mockZonesResponse got too many zone names")
 	}
 
-	zones := make([]cloudflare.Zone, numZones)
-	for i := 0; i < numZones; i++ {
-		zones[i] = *mockZone(zoneName, i)
+	zones := make([]cloudflare.Zone, len(zoneStatuses))
+	for i, status := range zoneStatuses {
+		zones[i] = *mockZone(zoneName, i, status)
 	}
 
 	return &cloudflare.ZonesResponse{
@@ -195,7 +198,7 @@ func mockZonesResponse(zoneName string, numZones int) *cloudflare.ZonesResponse 
 	}
 }
 
-func handleZones(t *testing.T, zoneName string, numZones int, w http.ResponseWriter, r *http.Request) {
+func handleZones(t *testing.T, zoneName string, zoneStatuses []string, w http.ResponseWriter, r *http.Request) {
 	t.Helper()
 
 	require.Equal(t, http.MethodGet, r.Method)
@@ -204,26 +207,25 @@ func handleZones(t *testing.T, zoneName string, numZones int, w http.ResponseWri
 		"account.id": {mockAccount},
 		"name":       {zoneName},
 		"per_page":   {"50"},
-		"status":     {"active"},
 	}, r.URL.Query())
 
 	w.Header().Set("content-type", "application/json")
-	err := json.NewEncoder(w).Encode(mockZonesResponse(zoneName, numZones))
+	err := json.NewEncoder(w).Encode(mockZonesResponse(zoneName, zoneStatuses))
 	require.NoError(t, err)
 }
 
 type zonesHandler struct {
-	mux         *http.ServeMux
-	numZones    *map[string]int
-	accessCount *int
+	mux          *http.ServeMux
+	zoneStatuses *map[string][]string
+	accessCount  *int
 }
 
 func newZonesHandler(t *testing.T, mux *http.ServeMux) *zonesHandler {
 	t.Helper()
 
 	var (
-		numZones    map[string]int
-		accessCount int
+		zoneStatuses map[string][]string
+		accessCount  int
 	)
 
 	mux.HandleFunc("/zones", func(w http.ResponseWriter, r *http.Request) {
@@ -233,18 +235,18 @@ func newZonesHandler(t *testing.T, mux *http.ServeMux) *zonesHandler {
 		accessCount--
 
 		zoneName := r.URL.Query().Get("name")
-		handleZones(t, zoneName, numZones[zoneName], w, r)
+		handleZones(t, zoneName, zoneStatuses[zoneName], w, r)
 	})
 
 	return &zonesHandler{
-		mux:         mux,
-		numZones:    &numZones,
-		accessCount: &accessCount,
+		mux:          mux,
+		zoneStatuses: &zoneStatuses,
+		accessCount:  &accessCount,
 	}
 }
 
-func (h *zonesHandler) set(numZones map[string]int, accessCount int) {
-	*(h.numZones), *(h.accessCount) = numZones, accessCount
+func (h *zonesHandler) set(zoneStatuses map[string][]string, accessCount int) {
+	*(h.zoneStatuses), *(h.accessCount) = zoneStatuses, accessCount
 }
 
 func (h *zonesHandler) isExhausted() bool {
@@ -271,7 +273,7 @@ func TestActiveZonesTwo(t *testing.T) {
 
 	zh := newZonesHandler(t, mux)
 
-	zh.set(map[string]int{"test.org": 2}, 1)
+	zh.set(map[string][]string{"test.org": {"active", "active"}}, 1)
 	mockPP := mocks.NewMockPP(mockCtrl)
 	zones, ok := h.(*api.CloudflareHandle).ActiveZones(context.Background(), mockPP, "test.org")
 	require.True(t, ok)
@@ -308,7 +310,7 @@ func TestActiveZonesEmpty(t *testing.T) {
 
 	zh := newZonesHandler(t, mux)
 
-	zh.set(map[string]int{}, 1)
+	zh.set(map[string][]string{}, 1)
 	mockPP := mocks.NewMockPP(mockCtrl)
 	zones, ok := h.(*api.CloudflareHandle).ActiveZones(context.Background(), mockPP, "test.org")
 	require.True(t, ok)
@@ -343,35 +345,35 @@ func TestZoneOfDomain(t *testing.T) {
 
 	for name, tc := range map[string]struct {
 		zone          string
-		domain        api.Domain
-		numZones      map[string]int
+		domain        domain.Domain
+		zoneStatuses  map[string][]string
 		accessCount   int
 		expected      string
 		ok            bool
 		prepareMockPP func(*mocks.MockPP)
 	}{
-		"root":     {"test.org", api.FQDN("test.org"), map[string]int{"test.org": 1}, 1, mockID("test.org", 0), true, nil},
-		"wildcard": {"test.org", api.Wildcard("test.org"), map[string]int{"test.org": 1}, 1, mockID("test.org", 0), true, nil}, //nolint:lll
-		"one":      {"test.org", api.FQDN("sub.test.org"), map[string]int{"test.org": 1}, 2, mockID("test.org", 0), true, nil}, //nolint:lll
+		"root":     {"test.org", domain.FQDN("test.org"), map[string][]string{"test.org": {"active"}}, 1, mockID("test.org", 0), true, nil},     //nolint:lll
+		"wildcard": {"test.org", domain.Wildcard("test.org"), map[string][]string{"test.org": {"active"}}, 1, mockID("test.org", 0), true, nil}, //nolint:lll
+		"one":      {"test.org", domain.FQDN("sub.test.org"), map[string][]string{"test.org": {"active"}}, 2, mockID("test.org", 0), true, nil}, //nolint:lll
 		"none": {
-			"test.org", api.FQDN("sub.test.org"),
-			map[string]int{},
+			"test.org", domain.FQDN("sub.test.org"),
+			map[string][]string{},
 			3, "", false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Warningf(pp.EmojiError, "Failed to find the zone of %q", "sub.test.org")
 			},
 		},
 		"none/wildcard": {
-			"test.org", api.Wildcard("test.org"),
-			map[string]int{},
+			"test.org", domain.Wildcard("test.org"),
+			map[string][]string{},
 			2, "", false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Warningf(pp.EmojiError, "Failed to find the zone of %q", "*.test.org")
 			},
 		},
 		"multiple": {
-			"test.org", api.FQDN("sub.test.org"),
-			map[string]int{"test.org": 2},
+			"test.org", domain.FQDN("sub.test.org"),
+			map[string][]string{"test.org": {"active", "active"}},
 			2, "", false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Warningf(
@@ -382,14 +384,58 @@ func TestZoneOfDomain(t *testing.T) {
 			},
 		},
 		"multiple/wildcard": {
-			"test.org", api.Wildcard("test.org"),
-			map[string]int{"test.org": 2},
+			"test.org", domain.Wildcard("test.org"),
+			map[string][]string{"test.org": {"active", "active"}},
 			1, "", false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Warningf(
 					pp.EmojiImpossible,
 					"Found multiple active zones named %q. Specifying CF_ACCOUNT_ID might help",
 					"test.org",
+				)
+			},
+		},
+		"deleted": {
+			"test.org", domain.FQDN("test.org"),
+			map[string][]string{"test.org": {"deleted"}},
+			2, "", false,
+			func(m *mocks.MockPP) {
+				gomock.InOrder(
+					m.EXPECT().Infof(pp.EmojiWarning, "Zone %q is %q and thus skipped", "test.org", "deleted"),
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to find the zone of %q", "test.org"),
+				)
+			},
+		},
+		"pending": {
+			"test.org", domain.FQDN("test.org"),
+			map[string][]string{"test.org": {"pending"}},
+			1, mockID("test.org", 0), true,
+			func(m *mocks.MockPP) {
+				gomock.InOrder(
+					m.EXPECT().Warningf(pp.EmojiWarning, "Zone %q is %q; your Cloudflare setup is incomplete", "test.org", "pending"), //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiWarning, "Some features might stop working", "test.org", "pending"),
+				)
+			},
+		},
+		"initializing": {
+			"test.org", domain.FQDN("test.org"),
+			map[string][]string{"test.org": {"initializing"}},
+			1, mockID("test.org", 0), true,
+			func(m *mocks.MockPP) {
+				gomock.InOrder(
+					m.EXPECT().Warningf(pp.EmojiWarning, "Zone %q is %q; your Cloudflare setup is incomplete", "test.org", "initializing"), //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiWarning, "Some features might stop working", "test.org", "initializing"),
+				)
+			},
+		},
+		"undocumented": {
+			"test.org", domain.FQDN("test.org"),
+			map[string][]string{"test.org": {"some-undocumented-status"}},
+			1, mockID("test.org", 0), true,
+			func(m *mocks.MockPP) {
+				gomock.InOrder(
+					m.EXPECT().Warningf(pp.EmojiImpossible, "Zone %q is in an undocumented status %q", "test.org", "some-undocumented-status"), //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiImpossible, "Please report the bug at https://github.com/favonia/cloudflare-ddns/issues/new"),  //nolint:lll
 				)
 			},
 		},
@@ -402,7 +448,7 @@ func TestZoneOfDomain(t *testing.T) {
 
 			zh := newZonesHandler(t, mux)
 
-			zh.set(tc.numZones, tc.accessCount)
+			zh.set(tc.zoneStatuses, tc.accessCount)
 			mockPP := mocks.NewMockPP(mockCtrl)
 			if tc.prepareMockPP != nil {
 				tc.prepareMockPP(mockPP)
@@ -414,10 +460,7 @@ func TestZoneOfDomain(t *testing.T) {
 
 			if tc.ok {
 				zh.set(nil, 0)
-				mockPP = mocks.NewMockPP(mockCtrl)
-				if tc.prepareMockPP != nil {
-					tc.prepareMockPP(mockPP)
-				}
+				mockPP = mocks.NewMockPP(mockCtrl) // there shouldn't be any messages
 				zoneID, ok = h.(*api.CloudflareHandle).ZoneOfDomain(context.Background(), mockPP, tc.domain)
 				require.Equal(t, tc.ok, ok)
 				require.Equal(t, tc.expected, zoneID)
@@ -440,7 +483,7 @@ func TestZoneOfDomainInvalid(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	zoneID, ok := h.(*api.CloudflareHandle).ZoneOfDomain(context.Background(), mockPP, api.FQDN("sub.test.org"))
+	zoneID, ok := h.(*api.CloudflareHandle).ZoneOfDomain(context.Background(), mockPP, domain.FQDN("sub.test.org"))
 	require.False(t, ok)
 	require.Equal(t, "", zoneID)
 }
@@ -505,7 +548,7 @@ func TestListRecords(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	var (
 		ipNet       ipnet.Type
@@ -536,14 +579,14 @@ func TestListRecords(t *testing.T) {
 	expected := map[string]netip.Addr{"record1": mustIP("::1"), "record2": mustIP("::2")}
 	ipNet, ips, accessCount = ipnet.IP6, expected, 1
 	mockPP := mocks.NewMockPP(mockCtrl)
-	ips, ok := h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	ips, ok := h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.True(t, ok)
 	require.Equal(t, expected, ips)
 	require.Equal(t, 0, accessCount)
 
 	// testing the caching
 	mockPP = mocks.NewMockPP(mockCtrl)
-	ips, ok = h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	ips, ok = h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.True(t, ok)
 	require.Equal(t, expected, ips)
 }
@@ -556,7 +599,7 @@ func TestListRecordsInvalidIPAddress(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	var (
 		ipNet       ipnet.Type
@@ -594,7 +637,7 @@ func TestListRecordsInvalidIPAddress(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	ips, ok := h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	ips, ok := h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.False(t, ok)
 	require.Nil(t, ips)
 	require.Equal(t, 0, accessCount)
@@ -607,7 +650,7 @@ func TestListRecordsInvalidIPAddress(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	ips, ok = h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	ips, ok = h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.False(t, ok)
 	require.Nil(t, ips)
 	require.Equal(t, 0, accessCount)
@@ -621,7 +664,7 @@ func TestListRecordsWildcard(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 1)
+	zh.set(map[string][]string{"test.org": {"active"}}, 1)
 
 	var (
 		ipNet       ipnet.Type
@@ -652,14 +695,14 @@ func TestListRecordsWildcard(t *testing.T) {
 	expected := map[string]netip.Addr{"record1": mustIP("::1"), "record2": mustIP("::2")}
 	ipNet, ips, accessCount = ipnet.IP6, expected, 1
 	mockPP := mocks.NewMockPP(mockCtrl)
-	ips, ok := h.ListRecords(context.Background(), mockPP, api.Wildcard("test.org"), ipnet.IP6)
+	ips, ok := h.ListRecords(context.Background(), mockPP, domain.Wildcard("test.org"), ipnet.IP6)
 	require.True(t, ok)
 	require.Equal(t, expected, ips)
 	require.Equal(t, 0, accessCount)
 
 	// testing the caching
 	mockPP = mocks.NewMockPP(mockCtrl)
-	ips, ok = h.ListRecords(context.Background(), mockPP, api.Wildcard("test.org"), ipnet.IP6)
+	ips, ok = h.ListRecords(context.Background(), mockPP, domain.Wildcard("test.org"), ipnet.IP6)
 	require.True(t, ok)
 	require.Equal(t, expected, ips)
 }
@@ -671,17 +714,17 @@ func TestListRecordsInvalidDomain(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	mockPP := mocks.NewMockPP(mockCtrl)
 	mockPP.EXPECT().Warningf(pp.EmojiError, "Failed to retrieve records of %q: %v", "sub.test.org", gomock.Any())
-	ips, ok := h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP4)
+	ips, ok := h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP4)
 	require.False(t, ok)
 	require.Nil(t, ips)
 
 	mockPP = mocks.NewMockPP(mockCtrl)
 	mockPP.EXPECT().Warningf(pp.EmojiError, "Failed to retrieve records of %q: %v", "sub.test.org", gomock.Any())
-	ips, ok = h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	ips, ok = h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.False(t, ok)
 	require.Nil(t, ips)
 }
@@ -699,7 +742,7 @@ func TestListRecordsInvalidZone(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	ips, ok := h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP4)
+	ips, ok := h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP4)
 	require.False(t, ok)
 	require.Nil(t, ips)
 
@@ -710,7 +753,7 @@ func TestListRecordsInvalidZone(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	ips, ok = h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	ips, ok = h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.False(t, ok)
 	require.Nil(t, ips)
 }
@@ -746,7 +789,7 @@ func TestDeleteRecordValid(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	var (
 		listAccessCount   int
@@ -784,14 +827,14 @@ func TestDeleteRecordValid(t *testing.T) {
 
 	deleteAccessCount = 1
 	mockPP := mocks.NewMockPP(mockCtrl)
-	ok := h.DeleteRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, "record1")
+	ok := h.DeleteRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, "record1")
 	require.True(t, ok)
 
 	listAccessCount, deleteAccessCount = 1, 1
 	mockPP = mocks.NewMockPP(mockCtrl)
-	_, _ = h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
-	_ = h.DeleteRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, "record1")
-	rs, ok := h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	_, _ = h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
+	_ = h.DeleteRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, "record1")
+	rs, ok := h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.True(t, ok)
 	require.Empty(t, rs)
 }
@@ -803,7 +846,7 @@ func TestDeleteRecordInvalid(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	mockPP := mocks.NewMockPP(mockCtrl)
 	mockPP.EXPECT().Warningf(pp.EmojiError, "Failed to delete a stale %s record of %q (ID: %s): %v",
@@ -812,7 +855,7 @@ func TestDeleteRecordInvalid(t *testing.T) {
 		"record1",
 		gomock.Any(),
 	)
-	ok := h.DeleteRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, "record1")
+	ok := h.DeleteRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, "record1")
 	require.False(t, ok)
 }
 
@@ -827,7 +870,7 @@ func TestDeleteRecordZoneInvalid(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	ok := h.DeleteRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, "record1")
+	ok := h.DeleteRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, "record1")
 	require.False(t, ok)
 }
 
@@ -839,7 +882,7 @@ func TestUpdateRecordValid(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	var (
 		listAccessCount   int
@@ -886,14 +929,14 @@ func TestUpdateRecordValid(t *testing.T) {
 
 	updateAccessCount = 1
 	mockPP := mocks.NewMockPP(mockCtrl)
-	ok := h.UpdateRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, "record1", mustIP("::2"))
+	ok := h.UpdateRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, "record1", mustIP("::2"))
 	require.True(t, ok)
 
 	listAccessCount, updateAccessCount = 1, 1
 	mockPP = mocks.NewMockPP(mockCtrl)
-	_, _ = h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
-	_ = h.UpdateRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, "record1", mustIP("::2"))
-	rs, ok := h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	_, _ = h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
+	_ = h.UpdateRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, "record1", mustIP("::2"))
+	rs, ok := h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.True(t, ok)
 	require.Equal(t, map[string]netip.Addr{"record1": mustIP("::2")}, rs)
 }
@@ -905,7 +948,7 @@ func TestUpdateRecordInvalid(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	mockPP := mocks.NewMockPP(mockCtrl)
 	mockPP.EXPECT().Warningf(pp.EmojiError, "Failed to update a stale %s record of %q (ID: %s): %v",
@@ -914,7 +957,7 @@ func TestUpdateRecordInvalid(t *testing.T) {
 		"record1",
 		gomock.Any(),
 	)
-	ok := h.UpdateRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, "record1", mustIP("::1"))
+	ok := h.UpdateRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, "record1", mustIP("::1"))
 	require.False(t, ok)
 }
 
@@ -929,7 +972,7 @@ func TestUpdateRecordInvalidZone(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	ok := h.UpdateRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, "record1", mustIP("::1"))
+	ok := h.UpdateRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, "record1", mustIP("::1"))
 	require.False(t, ok)
 }
 
@@ -941,7 +984,7 @@ func TestCreateRecordValid(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	var (
 		listAccessCount   int
@@ -989,15 +1032,15 @@ func TestCreateRecordValid(t *testing.T) {
 
 	createAccessCount = 1
 	mockPP := mocks.NewMockPP(mockCtrl)
-	actualID, ok := h.CreateRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, mustIP("::1"), 100, false) //nolint:lll
+	actualID, ok := h.CreateRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, mustIP("::1"), 100, false) //nolint:lll
 	require.True(t, ok)
 	require.Equal(t, "record1", actualID)
 
 	listAccessCount, createAccessCount = 1, 1
 	mockPP = mocks.NewMockPP(mockCtrl)
-	_, _ = h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
-	_, _ = h.CreateRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, mustIP("::1"), 100, false) //nolint:lll
-	rs, ok := h.ListRecords(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6)
+	_, _ = h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
+	_, _ = h.CreateRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, mustIP("::1"), 100, false) //nolint:lll
+	rs, ok := h.ListRecords(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6)
 	require.True(t, ok)
 	require.Equal(t, map[string]netip.Addr{"record1": mustIP("::1")}, rs)
 }
@@ -1009,7 +1052,7 @@ func TestCreateRecordInvalid(t *testing.T) {
 	mux, h := newHandle(t)
 
 	zh := newZonesHandler(t, mux)
-	zh.set(map[string]int{"test.org": 1}, 2)
+	zh.set(map[string][]string{"test.org": {"active"}}, 2)
 
 	mockPP := mocks.NewMockPP(mockCtrl)
 	mockPP.EXPECT().Warningf(pp.EmojiError, "Failed to add a new %s record of %q: %v",
@@ -1017,7 +1060,7 @@ func TestCreateRecordInvalid(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	actualID, ok := h.CreateRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, mustIP("::1"), 100, false) //nolint:lll
+	actualID, ok := h.CreateRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, mustIP("::1"), 100, false) //nolint:lll
 	require.False(t, ok)
 	require.Equal(t, "", actualID)
 }
@@ -1033,7 +1076,7 @@ func TestCreateRecordInvalidZone(t *testing.T) {
 		"sub.test.org",
 		gomock.Any(),
 	)
-	actualID, ok := h.CreateRecord(context.Background(), mockPP, api.FQDN("sub.test.org"), ipnet.IP6, mustIP("::1"), 100, false) //nolint:lll
+	actualID, ok := h.CreateRecord(context.Background(), mockPP, domain.FQDN("sub.test.org"), ipnet.IP6, mustIP("::1"), 100, false) //nolint:lll
 	require.False(t, ok)
 	require.Equal(t, "", actualID)
 }
