@@ -6,6 +6,7 @@ import (
 	"io"
 	"net/http"
 	"net/url"
+	"strconv"
 	"strings"
 	"time"
 
@@ -13,10 +14,9 @@ import (
 )
 
 type HealthChecks struct {
-	BaseURL         string
-	RedactedBaseURL string
-	Timeout         time.Duration
-	MaxRetries      int
+	BaseURL    *url.URL
+	Timeout    time.Duration
+	MaxRetries int
 }
 
 const (
@@ -38,21 +38,20 @@ func SetHealthChecksMaxRetries(maxRetries int) HealthChecksOption {
 func NewHealthChecks(ppfmt pp.PP, rawURL string, os ...HealthChecksOption) (Monitor, bool) {
 	url, err := url.Parse(rawURL)
 	if err != nil {
-		ppfmt.Errorf(pp.EmojiUserError, "Failed to parse the Healthchecks URL %q: %v", rawURL, err)
+		ppfmt.Errorf(pp.EmojiUserError, "Failed to parse the Healthchecks.io URL (redacted)")
 		return nil, false
 	}
 
-	if !(url.IsAbs() && url.Opaque == "" && url.Host != "" && url.Fragment == "" && !url.ForceQuery && url.RawQuery == "") { //nolint: lll
-		ppfmt.Errorf(pp.EmojiUserError, `The URL %q does not look like a valid Healthchecks URL.`, url.Redacted())
+	if !(url.IsAbs() && url.Opaque == "" && url.Host != "") {
+		ppfmt.Errorf(pp.EmojiUserError, `The Healthchecks.io URL (redacted) does not look like a valid URL.`)
 		ppfmt.Errorf(pp.EmojiUserError, `A valid example is "https://hc-ping.com/01234567-0123-0123-0123-0123456789abc".`)
 		return nil, false
 	}
 
 	h := &HealthChecks{
-		BaseURL:         url.String(),
-		RedactedBaseURL: url.Redacted(),
-		Timeout:         HealthChecksDefaultTimeout,
-		MaxRetries:      HealthChecksDefaultMaxRetries,
+		BaseURL:    url,
+		Timeout:    HealthChecksDefaultTimeout,
+		MaxRetries: HealthChecksDefaultMaxRetries,
 	}
 
 	for _, o := range os {
@@ -66,12 +65,15 @@ func (h *HealthChecks) DescribeService() string {
 	return "Healthchecks.io"
 }
 
-func (h *HealthChecks) DescribeBaseURL() string {
-	return h.RedactedBaseURL
-}
+//nolint:funlen
+func (h *HealthChecks) ping(ctx context.Context, ppfmt pp.PP, endpoint string) bool {
+	url := h.BaseURL.JoinPath(endpoint)
 
-//nolint: funlen
-func (h *HealthChecks) ping(ctx context.Context, ppfmt pp.PP, url string, redatedURL string) bool {
+	endpointDescription := "default (root)"
+	if endpoint != "" {
+		endpointDescription = strconv.Quote(endpoint)
+	}
+
 	for retries := 0; retries < h.MaxRetries; retries++ {
 		if retries > 0 {
 			time.Sleep(time.Second << (retries - 1))
@@ -80,22 +82,28 @@ func (h *HealthChecks) ping(ctx context.Context, ppfmt pp.PP, url string, redate
 		ctx, cancel := context.WithTimeout(ctx, h.Timeout)
 		defer cancel()
 
-		req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
+		req, err := http.NewRequestWithContext(ctx, http.MethodGet, url.String(), nil)
 		if err != nil {
-			ppfmt.Warningf(pp.EmojiImpossible, "Failed to prepare HTTP(S) request to %q: %v", redatedURL, err)
+			ppfmt.Warningf(pp.EmojiImpossible,
+				"Failed to prepare HTTP(S) request to the %s endpoint of Healthchecks.io: %v",
+				endpointDescription, err)
 			return false
 		}
 
 		resp, err := http.DefaultClient.Do(req)
 		if err != nil {
-			ppfmt.Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", redatedURL, err)
+			ppfmt.Warningf(pp.EmojiError,
+				"Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v",
+				endpointDescription, err)
 			ppfmt.Infof(pp.EmojiRepeatOnce, "Trying again . . .")
 			continue
 		}
 		defer resp.Body.Close()
 		body, err := io.ReadAll(resp.Body)
 		if err != nil {
-			ppfmt.Warningf(pp.EmojiError, "Failed to read HTTP(S) response from %q: %v", redatedURL, err)
+			ppfmt.Warningf(pp.EmojiError,
+				"Failed to read HTTP(S) response from the %s endpoint of Healthchecks.io: %v",
+				endpointDescription, err)
 			ppfmt.Infof(pp.EmojiRepeatOnce, "Trying again . . .")
 			continue
 		}
@@ -119,41 +127,41 @@ func (h *HealthChecks) ping(ctx context.Context, ppfmt pp.PP, url string, redate
 
 		bodyAsString := strings.TrimSpace(string(body))
 		if bodyAsString != "OK" {
-			ppfmt.Warningf(
-				pp.EmojiError,
-				"Failed to ping %q; got response code: %d %s",
-				redatedURL,
-				resp.StatusCode,
-				bodyAsString,
+			ppfmt.Warningf(pp.EmojiError,
+				"Failed to ping the %s endpoint of Healthchecks.io; got response code: %d %s",
+				endpointDescription, resp.StatusCode, bodyAsString,
 			)
 			return false
 		}
 
-		ppfmt.Infof(pp.EmojiNotification, "Successfully pinged %q.", redatedURL)
+		ppfmt.Infof(pp.EmojiNotification, "Successfully pinged the %s endpoint of Healthchecks.io", endpointDescription)
 		return true
 	}
 
-	ppfmt.Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q in %d time(s).", redatedURL, h.MaxRetries)
+	ppfmt.Warningf(
+		pp.EmojiError,
+		"Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io in %d time(s)",
+		endpointDescription, h.MaxRetries)
 	return false
 }
 
 func (h *HealthChecks) Success(ctx context.Context, ppfmt pp.PP) bool {
-	return h.ping(ctx, ppfmt, h.BaseURL, h.RedactedBaseURL)
+	return h.ping(ctx, ppfmt, "")
 }
 
 func (h *HealthChecks) Start(ctx context.Context, ppfmt pp.PP) bool {
-	return h.ping(ctx, ppfmt, h.BaseURL+"/start", h.RedactedBaseURL+"/start")
+	return h.ping(ctx, ppfmt, "/start")
 }
 
 func (h *HealthChecks) Failure(ctx context.Context, ppfmt pp.PP) bool {
-	return h.ping(ctx, ppfmt, h.BaseURL+"/fail", h.RedactedBaseURL+"/fail")
+	return h.ping(ctx, ppfmt, "/fail")
 }
 
 func (h *HealthChecks) ExitStatus(ctx context.Context, ppfmt pp.PP, code int) bool {
 	if code < 0 || code > 255 {
-		ppfmt.Errorf(pp.EmojiImpossible, "Exit code (%i) not within the range 0-255.", code)
+		ppfmt.Errorf(pp.EmojiImpossible, "Exit code (%i) not within the range 0-255", code)
 		return false
 	}
 
-	return h.ping(ctx, ppfmt, fmt.Sprintf("%s/%d", h.BaseURL, code), fmt.Sprintf("%s/%d", h.RedactedBaseURL, code))
+	return h.ping(ctx, ppfmt, fmt.Sprintf("/%d", code))
 }
