@@ -19,11 +19,11 @@ import (
 func TestSetHealthChecksMaxRetries(t *testing.T) {
 	t.Parallel()
 
-	m := &monitor.HealthChecks{} //nolint: exhaustruct
+	m := &monitor.HealthChecks{} //nolint:exhaustruct
 
 	monitor.SetHealthChecksMaxRetries(42)(m)
 
-	require.Equal(t, m, &monitor.HealthChecks{MaxRetries: 42}) //nolint: exhaustruct
+	require.Equal(t, m, &monitor.HealthChecks{MaxRetries: 42}) //nolint:exhaustruct
 }
 
 func TestSetHealthChecksMaxRetriesPanic(t *testing.T) {
@@ -44,14 +44,17 @@ func TestSetHealthChecksMaxRetriesPanic(t *testing.T) {
 func TestNewHealthChecks(t *testing.T) {
 	t.Parallel()
 
+	rawURL := "https://user:pass@host/path"
+	parsedURL, err := url.Parse(rawURL)
+	require.NoError(t, err)
+
 	mockCtrl := gomock.NewController(t)
 	mockPP := mocks.NewMockPP(mockCtrl)
-	m, ok := monitor.NewHealthChecks(mockPP, "https://user:pass@host/path", monitor.SetHealthChecksMaxRetries(100))
+	m, ok := monitor.NewHealthChecks(mockPP, rawURL, monitor.SetHealthChecksMaxRetries(100))
 	require.Equal(t, m, &monitor.HealthChecks{
-		BaseURL:         "https://user:pass@host/path",
-		RedactedBaseURL: "https://user:xxxxx@host/path",
-		Timeout:         monitor.HealthChecksDefaultTimeout,
-		MaxRetries:      100,
+		BaseURL:    parsedURL,
+		Timeout:    monitor.HealthChecksDefaultTimeout,
+		MaxRetries: 100,
 	})
 	require.True(t, ok)
 }
@@ -62,8 +65,8 @@ func TestNewHealthChecksFail1(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	mockPP := mocks.NewMockPP(mockCtrl)
 	gomock.InOrder(
-		mockPP.EXPECT().Errorf(pp.EmojiUserError, `The URL %q does not look like a valid Healthchecks URL.`, url.PathEscape("this is not a valid URL")), //nolint: lll
-		mockPP.EXPECT().Errorf(pp.EmojiUserError, `A valid example is "https://hc-ping.com/01234567-0123-0123-0123-0123456789abc".`),                    //nolint: lll
+		mockPP.EXPECT().Errorf(pp.EmojiUserError, `The Healthchecks.io URL (redacted) does not look like a valid URL.`),
+		mockPP.EXPECT().Errorf(pp.EmojiUserError, `A valid example is "https://hc-ping.com/01234567-0123-0123-0123-0123456789abc".`), //nolint:lll
 	)
 	_, ok := monitor.NewHealthChecks(mockPP, "this is not a valid URL")
 	require.False(t, ok)
@@ -74,7 +77,7 @@ func TestNewHealthChecksFail2(t *testing.T) {
 
 	mockCtrl := gomock.NewController(t)
 	mockPP := mocks.NewMockPP(mockCtrl)
-	mockPP.EXPECT().Errorf(pp.EmojiUserError, "Failed to parse the Healthchecks URL %q: %v", "://#?", gomock.Any())
+	mockPP.EXPECT().Errorf(pp.EmojiUserError, "Failed to parse the Healthchecks.io URL (redacted)")
 	_, ok := monitor.NewHealthChecks(mockPP, "://#?")
 	require.False(t, ok)
 }
@@ -89,17 +92,7 @@ func TestDescripbeService(t *testing.T) {
 	require.Equal(t, "Healthchecks.io", m.DescribeService())
 }
 
-func TestDescripbeBaseURL(t *testing.T) {
-	t.Parallel()
-
-	mockCtrl := gomock.NewController(t)
-	mockPP := mocks.NewMockPP(mockCtrl)
-	m, ok := monitor.NewHealthChecks(mockPP, "https://user:pass@host/path")
-	require.True(t, ok)
-	require.Equal(t, "https://user:xxxxx@host/path", m.DescribeBaseURL())
-}
-
-//nolint: funlen
+//nolint:funlen
 func TestEndPoints(t *testing.T) {
 	t.Parallel()
 
@@ -128,11 +121,11 @@ func TestEndPoints(t *testing.T) {
 			true, true,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged %q.", gomock.Any()),                            //nolint: lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `default (root)`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                                     //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `default (root)`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                                     //nolint:lll
+					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged the %s endpoint of Healthchecks.io", `default (root)`),                             //nolint:lll
 				)
 			},
 		},
@@ -145,11 +138,11 @@ func TestEndPoints(t *testing.T) {
 			false, false,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()),                 //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()),                 //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to ping %q; got response code: %d %s", gomock.Any(), 400, "invalid url format"), //nolint: lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `default (root)`, gomock.Any()),                 //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                                                     //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `default (root)`, gomock.Any()),                 //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                                                     //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to ping the %s endpoint of Healthchecks.io; got response code: %d %s", `default (root)`, 400, "invalid url format"), //nolint:lll
 				)
 			},
 		},
@@ -162,13 +155,13 @@ func TestEndPoints(t *testing.T) {
 			false, false,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q in %d time(s).", gomock.Any(), 3), //nolint: lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `default (root)`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                                     //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `default (root)`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                                     //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `default (root)`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                                     //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io in %d time(s)", `default (root)`, 3),  //nolint:lll
 				)
 			},
 		},
@@ -181,11 +174,11 @@ func TestEndPoints(t *testing.T) {
 			true, true,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged %q.", gomock.Any()),                            //nolint: lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `"/start"`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                               //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `"/start"`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                               //nolint:lll
+					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged the %s endpoint of Healthchecks.io", `"/start"`),                             //nolint:lll
 				)
 			},
 		},
@@ -198,11 +191,11 @@ func TestEndPoints(t *testing.T) {
 			true, true,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged %q.", gomock.Any()),                            //nolint: lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `"/fail"`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                              //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `"/fail"`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                              //nolint:lll
+					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged the %s endpoint of Healthchecks.io", `"/fail"`),                             //nolint:lll
 				)
 			},
 		},
@@ -215,11 +208,11 @@ func TestEndPoints(t *testing.T) {
 			true, true,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged %q.", gomock.Any()),                            //nolint: lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `"/0"`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                           //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `"/0"`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                           //nolint:lll
+					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged the %s endpoint of Healthchecks.io", `"/0"`),                             //nolint:lll
 				)
 			},
 		},
@@ -232,11 +225,11 @@ func TestEndPoints(t *testing.T) {
 			true, true,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to %q: %v", gomock.Any(), gomock.Any()), //nolint: lll
-					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                 //nolint: lll
-					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged %q.", gomock.Any()),                            //nolint: lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `"/1"`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                           //nolint:lll
+					m.EXPECT().Warningf(pp.EmojiError, "Failed to send HTTP(S) request to the %s endpoint of Healthchecks.io: %v", `"/1"`, gomock.Any()), //nolint:lll
+					m.EXPECT().Infof(pp.EmojiRepeatOnce, "Trying again . . ."),                                                                           //nolint:lll
+					m.EXPECT().Infof(pp.EmojiNotification, "Successfully pinged the %s endpoint of Healthchecks.io", `"/1"`),                             //nolint:lll
 				)
 			},
 		},
@@ -248,7 +241,7 @@ func TestEndPoints(t *testing.T) {
 			nil,
 			false, false,
 			func(m *mocks.MockPP) {
-				m.EXPECT().Errorf(pp.EmojiImpossible, "Exit code (%i) not within the range 0-255.", -1)
+				m.EXPECT().Errorf(pp.EmojiImpossible, "Exit code (%i) not within the range 0-255", -1)
 			},
 		},
 	} {
@@ -294,21 +287,4 @@ func TestEndPoints(t *testing.T) {
 			require.Equal(t, tc.pinged, pinged)
 		})
 	}
-}
-
-func TestEndPointsIllFormed(t *testing.T) {
-	t.Parallel()
-	mockCtrl := gomock.NewController(t)
-	mockPP := mocks.NewMockPP(mockCtrl)
-	mockPP.EXPECT().Warningf(pp.EmojiImpossible, "Failed to prepare HTTP(S) request to %q: %v", "blah", gomock.Any())
-
-	m := &monitor.HealthChecks{
-		BaseURL:         "://#?",
-		RedactedBaseURL: "blah",
-		Timeout:         monitor.HealthChecksDefaultTimeout,
-		MaxRetries:      monitor.HealthChecksDefaultMaxRetries,
-	}
-
-	ok := m.Success(context.Background(), mockPP)
-	require.False(t, ok)
 }
