@@ -5,6 +5,7 @@ import (
 	"net/netip"
 
 	"github.com/favonia/cloudflare-ddns/internal/api"
+	"github.com/favonia/cloudflare-ddns/internal/domain"
 	"github.com/favonia/cloudflare-ddns/internal/ipnet"
 	"github.com/favonia/cloudflare-ddns/internal/pp"
 )
@@ -12,5 +13,13 @@ import (
 //go:generate mockgen -destination=../mocks/mock_setter.go -package=mocks . Setter
 
 type Setter interface {
-	Set(ctx context.Context, ppfmt pp.PP, Domain api.Domain, IPNetwork ipnet.Type, IP netip.Addr) bool
+	Set(
+		ctx context.Context,
+		ppfmt pp.PP,
+		Domain domain.Domain,
+		IPNetwork ipnet.Type,
+		IP netip.Addr,
+		ttl api.TTL,
+		proxied bool,
+	) bool
 }
