@@ -9,26 +9,23 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/favonia/cloudflare-ddns/internal/api"
+	"github.com/favonia/cloudflare-ddns/internal/domain"
 	"github.com/favonia/cloudflare-ddns/internal/ipnet"
 	"github.com/favonia/cloudflare-ddns/internal/mocks"
 	"github.com/favonia/cloudflare-ddns/internal/pp"
 	"github.com/favonia/cloudflare-ddns/internal/setter"
 )
 
-//nolint:funlen,maintidx
+//nolint:funlen
 func TestSet(t *testing.T) {
 	t.Parallel()
 
-	type anys = []interface{}
-
 	const (
-		domain    = api.FQDN("sub.test.org")
+		domain    = domain.FQDN("sub.test.org")
 		ipNetwork = ipnet.IP6
 		record1   = "record1"
 		record2   = "record2"
 		record3   = "record3"
-		ttl       = api.TTL(100)
-		proxied   = true
 	)
 	var (
 		ip1       = netip.MustParseAddr("::1")
@@ -39,11 +36,15 @@ func TestSet(t *testing.T) {
 	for name, tc := range map[string]struct {
 		ip                netip.Addr
 		ok                bool
+		ttl               api.TTL
+		proxied           bool
 		prepareMockPP     func(m *mocks.MockPP)
 		prepareMockHandle func(ctx context.Context, ppfmt pp.PP, m *mocks.MockHandle)
 	}{
-		"0-nil": {
+		"0-nil/1-true": {
 			invalidIP,
+			true,
+			1,
 			true,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Infof(pp.EmojiAlreadyDone, "The %s records of %q are already up to date", "AAAA", "sub.test.org")
@@ -52,22 +53,26 @@ func TestSet(t *testing.T) {
 				m.EXPECT().ListRecords(ctx, ppfmt, domain, ipNetwork).Return(map[string]netip.Addr{}, true)
 			},
 		},
-		"0": {
+		"0/1-false": {
 			ip1,
 			true,
+			1,
+			false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Noticef(pp.EmojiAddRecord, "Added a new %s record of %q (ID: %s)", "AAAA", "sub.test.org", record1)
 			},
 			func(ctx context.Context, ppfmt pp.PP, m *mocks.MockHandle) {
 				gomock.InOrder(
 					m.EXPECT().ListRecords(ctx, ppfmt, domain, ipNetwork).Return(map[string]netip.Addr{}, true),
-					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, ttl, proxied).Return(record1, true),
+					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, api.TTL(1), false).Return(record1, true),
 				)
 			},
 		},
-		"1unmatched": {
+		"1unmatched/300-false": {
 			ip1,
 			true,
+			300,
+			false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Noticef(pp.EmojiUpdateRecord,
 					"Updated a stale %s record of %q (ID: %s)",
@@ -83,9 +88,11 @@ func TestSet(t *testing.T) {
 				)
 			},
 		},
-		"1unmatched-updatefail": {
+		"1unmatched-updatefail/300-false": {
 			ip1,
 			true,
+			300,
+			false,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
 					m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record1),
@@ -97,13 +104,15 @@ func TestSet(t *testing.T) {
 					m.EXPECT().ListRecords(ctx, ppfmt, domain, ipNetwork).Return(map[string]netip.Addr{record1: ip2}, true),
 					m.EXPECT().UpdateRecord(ctx, ppfmt, domain, ipNetwork, record1, ip1).Return(false),
 					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record1).Return(true),
-					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, ttl, proxied).Return(record2, true),
+					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, api.TTL(300), false).Return(record2, true),
 				)
 			},
 		},
-		"1unmatched-nil": {
+		"1unmatched-nil/300-false": {
 			invalidIP,
 			true,
+			300,
+			false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record1)
 			},
@@ -114,9 +123,11 @@ func TestSet(t *testing.T) {
 				)
 			},
 		},
-		"1matched": {
+		"1matched/300-false": {
 			ip1,
 			true,
+			300,
+			false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Infof(pp.EmojiAlreadyDone, "The %s records of %q are already up to date", "AAAA", "sub.test.org")
 			},
@@ -124,9 +135,11 @@ func TestSet(t *testing.T) {
 				m.EXPECT().ListRecords(ctx, ppfmt, domain, ipNetwork).Return(map[string]netip.Addr{record1: ip1}, true)
 			},
 		},
-		"2matched": {
+		"2matched/300-false": {
 			ip1,
 			true,
+			300,
+			false,
 			func(m *mocks.MockPP) {
 				m.EXPECT().Noticef(
 					pp.EmojiDelRecord,
@@ -143,9 +156,11 @@ func TestSet(t *testing.T) {
 				)
 			},
 		},
-		"2matched-deletefail": {
+		"2matched-deletefail/300-false": {
 			ip1,
 			true,
+			300,
+			false,
 			nil,
 			func(ctx context.Context, ppfmt pp.PP, m *mocks.MockHandle) {
 				gomock.InOrder(
@@ -154,9 +169,11 @@ func TestSet(t *testing.T) {
 				)
 			},
 		},
-		"2unmatched": {
+		"2unmatched/300-false": {
 			ip1,
 			true,
+			300,
+			false,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
 					m.EXPECT().Noticef(
@@ -177,9 +194,11 @@ func TestSet(t *testing.T) {
 				)
 			},
 		},
-		"2unmatched-updatefail": {
+		"2unmatched-updatefail/300-false": {
 			ip1,
 			true,
+			300,
+			false,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
 					m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record1),
@@ -201,9 +220,11 @@ func TestSet(t *testing.T) {
 				)
 			},
 		},
-		"2unmatched-updatefailtwice": {
+		"2unmatched-updatefailtwice/300-false": {
 			ip1,
 			true,
+			300,
+			false,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
 					m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record1),
@@ -218,19 +239,21 @@ func TestSet(t *testing.T) {
 					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record1).Return(true),
 					m.EXPECT().UpdateRecord(ctx, ppfmt, domain, ipNetwork, record2, ip1).Return(false),
 					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record2).Return(true),
-					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, ttl, proxied).Return(record3, true),
+					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, api.TTL(300), false).Return(record3, true),
 				)
 			},
 		},
 		//nolint:dupl
-		"2unmatched-updatefail-deletefail-updatefail": {
+		"2unmatched-updatefail-deletefail-updatefail/300-false": {
 			ip1,
+			false,
+			300,
 			false,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
 					m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record2),
 					m.EXPECT().Noticef(pp.EmojiAddRecord, "Added a new %s record of %q (ID: %s)", "AAAA", "sub.test.org", record3),
-					m.EXPECT().Errorf(pp.EmojiError, "Failed to (fully) update %s records of %q", "AAAA", "sub.test.org"),
+					m.EXPECT().Errorf(pp.EmojiError, "Failed to complete updating of %s records of %q; records might be inconsistent", "AAAA", "sub.test.org"), //nolint:lll
 				)
 			},
 			func(ctx context.Context, ppfmt pp.PP, m *mocks.MockHandle) {
@@ -240,19 +263,21 @@ func TestSet(t *testing.T) {
 					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record1).Return(false),
 					m.EXPECT().UpdateRecord(ctx, ppfmt, domain, ipNetwork, record2, ip1).Return(false),
 					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record2).Return(true),
-					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, ttl, proxied).Return(record3, true),
+					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, api.TTL(300), false).Return(record3, true),
 				)
 			},
 		},
 		//nolint:dupl
-		"2unmatched-updatefailtwice-createfail": {
+		"2unmatched-updatefailtwice-createfail/300-false": {
 			ip1,
+			false,
+			300,
 			false,
 			func(m *mocks.MockPP) {
 				gomock.InOrder(
 					m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record1),
 					m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record2),
-					m.EXPECT().Errorf(pp.EmojiError, "Failed to (fully) update %s records of %q", "AAAA", "sub.test.org"),
+					m.EXPECT().Errorf(pp.EmojiError, "Failed to complete updating of %s records of %q; records might be inconsistent", "AAAA", "sub.test.org"), //nolint:lll
 				)
 			},
 			func(ctx context.Context, ppfmt pp.PP, m *mocks.MockHandle) {
@@ -262,18 +287,40 @@ func TestSet(t *testing.T) {
 					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record1).Return(true),
 					m.EXPECT().UpdateRecord(ctx, ppfmt, domain, ipNetwork, record2, ip1).Return(false),
 					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record2).Return(true),
-					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, ttl, proxied).Return(record3, false),
+					m.EXPECT().CreateRecord(ctx, ppfmt, domain, ipNetwork, ip1, api.TTL(300), false).Return(record3, false),
 				)
 			},
 		},
-		"listfail": {
+		"listfail/300-false": {
 			ip1,
 			false,
+			300,
+			false,
 			func(m *mocks.MockPP) {
-				m.EXPECT().Errorf(pp.EmojiError, "Failed to (fully) update %s records of %q", "AAAA", "sub.test.org")
+				m.EXPECT().Errorf(pp.EmojiError, "Failed to retrieve the current %s records of %q", "AAAA", "sub.test.org")
 			},
 			func(ctx context.Context, ppfmt pp.PP, m *mocks.MockHandle) {
 				m.EXPECT().ListRecords(ctx, ppfmt, domain, ipNetwork).Return(nil, false)
+			},
+		},
+		"impossible-invalidIP-records/300-false": {
+			invalidIP,
+			true,
+			300,
+			false,
+			func(m *mocks.MockPP) {
+				gomock.InOrder(
+					m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record1),
+					m.EXPECT().Noticef(pp.EmojiDelRecord, "Deleted a stale %s record of %q (ID: %s)", "AAAA", "sub.test.org", record2),
+				)
+			},
+			func(ctx context.Context, ppfmt pp.PP, m *mocks.MockHandle) {
+				gomock.InOrder(
+					m.EXPECT().ListRecords(ctx, ppfmt, domain, ipNetwork).Return(map[string]netip.Addr{record1: ip1, record2: invalidIP}, true), //nolint:lll
+
+					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record1).Return(true),
+					m.EXPECT().DeleteRecord(ctx, ppfmt, domain, ipNetwork, record2).Return(true),
+				)
 			},
 		},
 	} {
@@ -293,10 +340,10 @@ func TestSet(t *testing.T) {
 				tc.prepareMockHandle(ctx, mockPP, mockHandle)
 			}
 
-			s, ok := setter.New(mockPP, mockHandle, ttl, proxied)
+			s, ok := setter.New(mockPP, mockHandle)
 			require.True(t, ok)
 
-			ok = s.Set(ctx, mockPP, domain, ipNetwork, tc.ip)
+			ok = s.Set(ctx, mockPP, domain, ipNetwork, tc.ip, tc.ttl, tc.proxied)
 			require.Equal(t, tc.ok, ok)
 		})
 	}
